requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pypiwin32==223
pytz==2022.2.1
pywin32==304
requests==2.28.1
urllib3==1.26.12
win10toast==0.9
yarl==1.8.1
